テストのエラー構文エラーを解決
