+++ app/backend/test/ScoreHistoryApi.Tests.WithFake/Logics/Scores/ScoreCreatorTests.cs
@@ -1,0 +1,90 @@
+using System;
+using System.Collections.Generic;
+using System.Threading.Tasks;
+using Amazon.DynamoDBv2;
+using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.DependencyInjection;
+using ScoreHistoryApi.Factories;
+using ScoreHistoryApi.Logics;
+using ScoreHistoryApi.Logics.ScoreDatabases;
+using ScoreHistoryApi.Logics.Scores;
+using Xunit;
+
+namespace ScoreHistoryApi.Tests.WithFake.Logics.Scores
+{
+    public class ScoreCreatorTests
+    {
+        public const string ScoreTableName = "ura-kata-score-history";
+        public const string ScoreBucket = "ura-kata-score-history-bucket";
+
+        [Fact]
+        public async Task CreateAsyncTest()
+        {
+            var factory = new DynamoDbClientFactory().SetEndpointUrl(new Uri("http://localhost:18000"));
+            var configuration = new ConfigurationBuilder()
+                .AddInMemoryCollection(new Dictionary<string, string>()
+                {
+                    [EnvironmentNames.ScoreDynamoDbTableName] = ScoreTableName,
+                    [EnvironmentNames.ScoreLargeDataDynamoDbTableName] = ScoreTableName,
+                    [EnvironmentNames.ScoreItemDynamoDbTableName] = ScoreTableName,
+                    [EnvironmentNames.ScoreItemRelationDynamoDbTableName] = ScoreTableName,
+                    [EnvironmentNames.ScoreItemS3Bucket] = ScoreBucket,
+                    [EnvironmentNames.ScoreDataSnapshotS3Bucket] = ScoreBucket,
+                })
+                .Build();
+            var serviceCollection = new ServiceCollection();
+
+            serviceCollection.AddSingleton(factory.Create());
+            serviceCollection.AddSingleton<ScoreQuota>();
+            serviceCollection.AddSingleton<IConfiguration>(configuration);
+            serviceCollection.AddScoped<Initializer>();
+            serviceCollection.AddScoped<ScoreDeleter>();
+            serviceCollection.AddScoped<ScoreDetailGetter>();
+
+            serviceCollection.AddScoped<ScoreCreator>();
+
+            await using var provider = serviceCollection.BuildServiceProvider();
+
+
+            var initializer = provider.GetRequiredService<Initializer>();
+            var deleter = provider.GetRequiredService<ScoreDeleter>();
+            var getter = provider.GetRequiredService<ScoreDetailGetter>();
+
+            var target = provider.GetRequiredService<ScoreCreator>();
+
+            var ownerId = Guid.Parse("eb184d71-3b6e-4619-a1f6-1ddb41de72f0");
+            var scoreId = Guid.Parse("0d9fb491-43ae-44a1-9056-55bb25b21187");
+            try
+            {
+                await initializer.InitializeScoreAsync(ownerId);
+            }
+            catch
+            {
+                // 初期化のエラーは握りつぶす
+            }
+
+            try
+            {
+                await deleter.DeleteScoreAsync(ownerId, scoreId);
+            }
+            catch
+            {
+                // 初期化のエラーは握りつぶす
+            }
+
+            var title = "test score";
+            var description = "楽譜の説明";
+
+            await target.CreateAsync(ownerId, scoreId, title, description);
+
+            var (scoreData, hashSet) = await getter.GetDynamoDbScoreDataAsync(ownerId, scoreId);
+
+            Assert.IsType<DynamoDbScoreDataV1>(scoreData.Data);
+
+            var dataV1 = (DynamoDbScoreDataV1) scoreData.Data;
+
+            Assert.Equal(title, dataV1.Title);
+            Assert.Equal(description, hashSet[dataV1.DescriptionHash]);
+        }
+    }
+}
