+++ app/backend/test/ScoreHistoryApi.Tests.WithFake/Logics/Scores/ScoreTitleSetterTests.cs
@@ -1,0 +1,80 @@
+using System;
+using System.Collections.Generic;
+using System.Threading.Tasks;
+using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.DependencyInjection;
+using ScoreHistoryApi.Factories;
+using ScoreHistoryApi.Logics;
+using ScoreHistoryApi.Logics.Scores;
+using Xunit;
+
+namespace ScoreHistoryApi.Tests.WithFake.Logics.Scores
+{
+    public class ScoreTitleSetterTests
+    {
+        public const string ScoreTableName = "ura-kata-score-history";
+        public const string ScoreBucket = "ura-kata-score-history-bucket";
+
+        [Fact]
+        public async Task UpdateTitleAsyncTest()
+        {
+            var factory = new DynamoDbClientFactory().SetEndpointUrl(new Uri("http://localhost:18000"));
+            var configuration = new ConfigurationBuilder()
+                .AddInMemoryCollection(new Dictionary<string, string>()
+                {
+                    [EnvironmentNames.ScoreDynamoDbTableName] = ScoreTableName,
+                    [EnvironmentNames.ScoreLargeDataDynamoDbTableName] = ScoreTableName,
+                    [EnvironmentNames.ScoreItemDynamoDbTableName] = ScoreTableName,
+                    [EnvironmentNames.ScoreItemRelationDynamoDbTableName] = ScoreTableName,
+                    [EnvironmentNames.ScoreItemS3Bucket] = ScoreBucket,
+                    [EnvironmentNames.ScoreDataSnapshotS3Bucket] = ScoreBucket,
+                })
+                .Build();
+            var serviceCollection = new ServiceCollection();
+
+            serviceCollection.AddSingleton(factory.Create());
+            serviceCollection.AddSingleton<ScoreQuota>();
+            serviceCollection.AddSingleton<IConfiguration>(configuration);
+            serviceCollection.AddScoped<Initializer>();
+            serviceCollection.AddScoped<ScoreDeleter>();
+            serviceCollection.AddScoped<ScoreDetailGetter>();
+
+            serviceCollection.AddScoped<ScoreTitleSetter>();
+
+            await using var provider = serviceCollection.BuildServiceProvider();
+
+
+            var initializer = provider.GetRequiredService<Initializer>();
+            var creator = provider.GetRequiredService<ScoreCreator>();
+
+
+            var target = provider.GetRequiredService<ScoreTitleSetter>();
+
+            var ownerId = Guid.Parse("27f30e40-bf83-427f-b53d-39b5d0f01ac7");
+            var scoreId = Guid.Parse("4afded99-4070-4ba7-85ed-6c3776602895");
+
+            var title = "test score";
+            var description = "楽譜の説明";
+
+            try
+            {
+                await initializer.InitializeScoreAsync(ownerId);
+            }
+            catch
+            {
+                // 初期化のエラーは握りつぶす
+            }
+            try
+            {
+                await creator.CreateAsync(ownerId, scoreId, title, description);
+            }
+            catch
+            {
+                // 初期化のエラーは握りつぶす
+            }
+
+            var newTitle = "new test score";
+            await target.UpdateTitleAsync(ownerId, scoreId, newTitle);
+        }
+    }
+}
